Add files via upload
Write a java program to create Person class with variables id, and name. Include methods 
getPerson() and showPerson() in the class. Again, create another class named Employee, 
which is child of person class. Include member variable salary, and methods 
getEmployee(), and showEmployee() in the class. Again, create a third class named Typist, 
which is child of Employee. Include variable typespeed, and methods getTypist(), and 
showTypist() in the class. Finally, create two objects of Typist class and read and display 
their details.